No need to validate href == idConc.
I think removing this check would be good as it will allow <button> or <input> elements to also open up the popup. This check restricts the usage to <a> elements as it needs a 'href' attribute.

Also, as the developer would have already specified the modal with the 'modalTarget' option, why do we need a cross-check again with the href element? I do not see any particular use for it.

Please let me know if I am missing out on anything particularly important related to the check.